init and run changed to uint16_t

+++ moto/src/moto_interface.h
@@ -43,6 +43,6 @@ extern Servo escRear;
  */
 static int moto_cyclesSinceLastMsg;
 
-uint8_t moto_init(void);
-uint8_t moto_run(void);
+uint16_t moto_init(void);
+uint16_t moto_run(void);
 #endif
